split autogen examples into two files and added a teacher-student example
