+++ samples/csharp/autogen/autogen.console.csproj
@@ -11,6 +11,7 @@
       <PackageReference Include="Strathweb.Phi.Engine" Version="0.1.0" />
       <PackageReference Include="AutoGen" Version="0.1.0" />
       <PackageReference Include="FluentAssertions" Version="6.12.0" />
-        <PackageReference Include="System.Text.Json" Version="8.0.4" />
+      <PackageReference Include="System.Text.Json" Version="8.0.4" />
+      <PackageReference Include="spectre.console" Version="0.47.0" />
   </ItemGroup>
 </Project>
